Add in dockerized versions of p4d and the swarm server, so that integration tests can work off of a fixed state.

+++ integration-test/src/docker/start-swarm-cluster.sh
@@ -33,8 +33,18 @@ if [ -z "$P4D_VERSION" ]; then
     P4D_VERSION=r18.1
 fi
 
+# TODO download a p4 executable.
+
 cd $(dirname "$0") || exit 1
 cp ../../../p4java/src/test/resources/bin/$P4D_VERSION/bin.$P4D_PLATFORM/p4d p4d/p4d.bin || exit 1
+if [ ! -f p4d/sampledepot.tar.gz ]; then
+    curl -L -o p4d/sampledepot.tar.gz http://ftp.perforce.com/pub/perforce/tools/sampledepot.tar.gz || exit 1
+fi
+if [ ! -f p4.bin ]; then
+    curl -L -o p4.bin http://ftp.perforce.com/perforce/r18.2/bin.linux26x86_64/p4 || exit 1
+fi
+cp p4.bin p4d/.
+cp p4.bin swarm/.
 
 if [ ! -f "swarm-$SWARM_VERSION.tgz" ]; then
     curl -L -o "swarm-$SWARM_VERSION.tgz" "$SWARM_URL" || exit 1
@@ -42,5 +52,6 @@ fi
 test -f swarm/swarm.tgz && rm swarm/swarm.tgz
 cp "$(pwd)/swarm-$SWARM_VERSION.tgz" swarm/swarm.tgz
 
-( cd p4d && docker build -t local/p4d . ) || exit 1
-( cd swarm && docker build -t local/swarm . ) || exit 1
+#( cd p4d && docker build -t local/p4d . ) || exit 1
+#( cd swarm && docker build -t local/swarm . ) || exit 1
+exec docker-compose up --build
